requirements.txt
# AI Employee — Dependencies
# Install with: .venv/bin/pip install -r requirements.txt

# ── Bronze Tier ───────────────────────────────────────────────
# File system monitoring
watchdog>=4.0.0

# ── Silver Tier ───────────────────────────────────────────────
# Gmail API (OAuth2 + Google API client)
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0

# Playwright Python package (browser automation for LinkedIn)
playwright>=1.58.0
# After installing, run once: .venv/bin/playwright install chromium
